pyproject.toml
[project]
name = "erpnext-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "motor (>=3.7.0,<4.0.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
